WilsonMG: Save even more calls to blockProject in CoarsenOperator
With this commit, the number of calls for a setup with nbasis = 48 is reduced from 864 to 240 which theoretically yields
a speedup of 3.6 in the time spent in blockProject (= the bottleneck). The output coarse link fields were checked
against the ones constructed with the original coarse operator construction and the results agreed up to a relative
deviation of ~ 1e-30.